refactor: 패키지 이동, 코드 중복 제거
