+++ src/test/java/wooteco/subway/acceptance/line/LineAcceptanceTest.java
@@ -2,10 +2,9 @@ package wooteco.subway.acceptance.line;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-
+import wooteco.subway.acceptance.AcceptanceTest;
 import wooteco.subway.service.line.dto.LineDetailResponse;
 import wooteco.subway.service.line.dto.LineResponse;
-import wooteco.subway.AcceptanceTest;
 
 import java.time.LocalTime;
 import java.util.List;
