+++ src/test/java/wooteco/subway/acceptance/member/LoginAcceptanceTest.java
@@ -1,18 +1,17 @@
 package wooteco.subway.acceptance.member;
 
-import static org.assertj.core.api.Assertions.*;
-
-import java.util.HashMap;
-import java.util.Map;
-
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
-
-import wooteco.subway.AuthAcceptanceTest;
+import wooteco.subway.acceptance.AuthAcceptanceTest;
 import wooteco.subway.service.member.dto.MemberResponse;
 import wooteco.subway.service.member.dto.TokenResponse;
+
+import java.util.HashMap;
+import java.util.Map;
+
+import static org.assertj.core.api.Assertions.assertThat;
 
 public class LoginAcceptanceTest extends AuthAcceptanceTest {
